GetPreguntas en Examen y Materia service

+++ Gex.NetCore/Services/Interface/IMateriaService.cs
@@ -8,8 +8,8 @@ namespace Gex.Services.Interface;
 public interface IMateriaService
 {
     Task<GexResult<ICollection<MateriaResponse>>> GetMateriasAsync();
-    Task<GexResult<MateriaResponse>> GetMateriaAsync(int id);
+    Task<GexResult<MateriaResponse>> GetMateriaAsync(long id);
     Task<GexResult<MateriaResponse>> CreateMateriaAsync(MateriaRequest materiaDto);
     Task<GexResult<MateriaResponse>> UpdateMateriaAsync(MateriaRequest materiaDto);
-    Task<GexResult<MateriaResponse>> DeleteMateriaAsync(int id);
+    Task<GexResult<MateriaResponse>> DeleteMateriaAsync(long id);
 }
